requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
pytest==8.3.4
